// src/services/api/auth/type.ts
export type authLoginTypeProps = {
    // email: string
    // password: string
    address: string
    network: string
    chainId: number
    balance: number
    signature: string
}

export type signUpUsernameTypeProps = {
    name: string
    email: string
    password: string
    role: string
    income_option: number
    income_option_select: number
    mobilephone_number: string
    phone_number: string
    zip1: string
    zip2: string
    address: string
    address1: string
    addressdoro: string
    zonecode: string
    centerId?: number
    recomid: string
    sponid: string
    return_bank: string
    return_account: string
    return_name: string
}